requirements.txt
nextcord[voice]
pyttsx3
Pillow
ffmpeg
